requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-jose
passlib
SQLAlchemy
psycopg2
alembic